Changed so that the logic that switches between Cert View and Co2 post is no longer part of CertManager but an external comp.
